Time: 17 ms (75.13%), Space: 82.8 MB (35.64%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.java
@@ -1,6 +1,6 @@
 class Solution {
     public int minDays(int[] bloomDay, int m, int k) {
-        int val=m*k;
+        long val=(long)m*k;
         if(val>bloomDay.length)return -1;
         int s=min(bloomDay);
         int e=max(bloomDay);
